Get human cancer predisposition genes

Downloads and retrieves a pre-processed dataset of human cancer predisposition genes. Genes deemed relevant for cancer predisposition have been collected from multiple resources:

Genomics England PanelApp

TCGA's PanCancer study

Cancer Gene Census

Human DNA repair genes

Other/curated

The dataset comes as a list object, with two elements:

metadata - a data frame with metadata regarding annotation resources used

records - a list with records of cancer predisposing gene, one record per gene

Usage

get_predisposition(cache_dir = NA, overwrite = F)

Arguments

cache_dir

Local directory for data download

overwrite

Logical indicating if local cache should be overwritten

Value

metadata - A data frame with 1 row and 6 columns:

source - gene annotation source

annotation_data - type of annotations used

url - URL of annotation resource

citation - publication to cite for annotation source (citation; PMID)

version - version used

abbreviation - abbreviation used in column names of records

records - A data frame with 591 rows and 8 columns:

symbol - official gene symbol

entrezgene - Entrez gene identifier

moi - mechanism of inheritance (AD, AR)

predisp_syndrome_cui - Concept unique identifiers (CUI, UMLS) - inherited cancer syndromes

predisp_cancer_cui - Concept unique identifiers (CUI, UMLS) - inherited cancer conditions

predisp_source - Sources for cancer-predisposition gene PANEL_APP, CGC, TCGA_PANCAN_2018, OTHER, ACMG_SF

phenotypes - associated cancer phenotypes

mechanism_of_disease - mechanism of disease)

Details

NOTE: The dataset also contains genes recommended for reporting of incidental findings (ACMG_SF).